+++ src/rule-nlp/SubMatch.hpp
@@ -9,7 +9,7 @@ typedef struct SubMatch {
     size_t end;
     std::string content;
     std::string category;
-    long long score;
+    int64_t score;
 } SubMatch;
 
 bool operator<(const SubMatch& lhs, const SubMatch& rhs);
